Move the options to the s3options.scala
