+++ src/main/scala/zio/s3/Test.scala
@@ -126,7 +126,7 @@ object Test {
         bucketName: String,
         key: String,
         content: ZStream[R, Throwable, Byte],
-        options: MultipartUploadOptions = MultipartUploadOptions.default
+        options: MultipartUploadOptions = MultipartUploadOptions()
       ): ZIO[R, S3Exception, Unit] =
         putObject(bucketName, key, 0, content.chunkN(options.partSize), options.uploadOptions)
     }
